Work spaces?
TODO:
Add work space detection when a new folder is added to the global work space. But where can we get it?
Stop the language crashing when opening a file does not exist in work space. Currently we have to keep it crash to refresh work space information when a new folder is added into work space